refactor: Redirect 페이지의 비즈니스 로직 추상화
